Keep and check types for objects transferred between C++ and Lua.
Keep type information for objects transferred from C++ to Lua. Check types
transferring back to C++.

Crossing the border from C++ to Lua keeps type information for instances
of registered classes:
 - as parameter to Lua functions
 - as return value of C++ callbacks

Obtaining references or pointers to instances of registered classes from
Lua fails if the actual type is not the requested:
 - as parameter to C++ callbacks (raises an error in Lua)
 - as return value of Lua functions
   - for pointers (returns nullptr)
   - for references (throws sel::TypeError)

Move datastructures of sel::MetatableRegistry to Lua. This reduces the
need to pass the MetatableRegistry through several layers down to the few
places it is really used.